renamed component files to kebab-case format and changed extension to from .tsx to .js
