+++ src/components/app-header/app-header.js
@@ -1,9 +1,9 @@
 import React from "react";
 import styles from './app-header.module.css';
 import { ProfileIcon } from '@ya.praktikum/react-developer-burger-ui-components';
-import NavLink from './nav-link/NavLink';
-import NavBar from './nav-bar/NavBar';
-import AppLogo from './app-logo/AppLogo';
+import NavLink from './nav-link/nav-link';
+import NavBar from './nav-bar/nav-bar';
+import AppLogo from './app-logo/app-logo';
 
 function AppHeader() {
     return (
