+++ src/components/app/app.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import AppHeader from '../app-header/AppHeader';
-import Main from '../main/Main';
+import AppHeader from '../app-header/app-header';
+import Main from '../main/main';
 
 function App() {
   return (
