+++ src/components/burger-constructor/burger-component/burger-component.js
