+++ src/components/burger-ingredients/tab-bar/tab-bar.js
