+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/app/App';
+import App from './components/app/app';
 // import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
